requirements.txt
atomicwrites==1.4.0
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
#zipp==3.1.0
